requirements.txt
pytest
torch
torchvision
pyyaml
wandb
numpy
Pillow